+++ src/lib/components/NavigationContextMenu.svelte
@@ -95,20 +95,11 @@
             <button
               type="button"
               role="menuitem"
-              data-menu-key="n"
+              data-menu-key="m"
               class:active-menu-item={folderSortFor(contextMenu.node.path) === "modified-desc"}
               on:click={() => handleContextMenuAction("sort:modified-desc")}
             >
-              Modified: <span class="menu-mnemonic">n</span>ewest first
-            </button>
-            <button
-              type="button"
-              role="menuitem"
-              data-menu-key="o"
-              class:active-menu-item={folderSortFor(contextMenu.node.path) === "modified-asc"}
-              on:click={() => handleContextMenuAction("sort:modified-asc")}
-            >
-              Modified: <span class="menu-mnemonic">o</span>ldest first
+              Recently <span class="menu-mnemonic">m</span>odified
             </button>
             <button
               type="button"